+++ packages/bluesky/lib/bluesky.dart
@@ -213,6 +213,7 @@ export 'package:bluesky/src/entities/facet.dart';
 export 'package:bluesky/src/entities/facet_feature.dart';
 export 'package:bluesky/src/entities/facet_link.dart';
 export 'package:bluesky/src/entities/facet_mention.dart';
+export 'package:bluesky/src/entities/facet_tag.dart';
 export 'package:bluesky/src/entities/feed.dart';
 export 'package:bluesky/src/entities/feed_generator.dart';
 export 'package:bluesky/src/entities/feed_generator_info.dart';