+++ packages/bluesky_text/lib/src/regex.dart
@@ -6,3 +6,4 @@ final handleRegex = RegExp(r'(^|\s|\()(@)([a-zA-Z0-9.-]+)(\b)');
 final linkRegex = RegExp(
   r'(^|\s|\()((https?:\/\/[\S]+)|((?<domain>[a-z][a-z0-9]*(\.[a-z0-9]+)+)[\S]*))',
 );
+final hashtagRegex = RegExp(r'(?:^|\s)(#[^\d\s]\S*)(?=\s)?');